feat(back): Setting up uid to user (!13)
Right now the user is saved with a "email" as uid that is not a working email.

This merge request removes the fake email and adds a real feide mail and gives each user a uid.


Make sure to test the frontend fully before approve'ing a review

Co-authored-by: Sondre Alfnes <sondre.alfnes@gmail.com>

Closes #20

See merge request sondralf/itp2-reflection-tool!13

* fix(back): made the get course endpoint also return the uid of the users

* pipeline

* fix(frontend): UID implementation

* feat(back): adding ids

* feat(db & back): Update to get user id

Co-authored-by: Julian <julianao@stud.ntnu.no>
Co-authored-by: Sondre Alfnes <sondre.alfnes@gmail.com>