ref(back): Refactor file structure and adds tests (!18)
How to test:

Run the backend by using the new command `make dev` as the Readme says.

Test the backend by running `make test` file.

To test this with an windows computer, check out Makefile and run the lines individually


Description:

Refactors the backend and adds testing, and adds a new way of running backend.

It is bad practice to have every file open in backend folder and not structured

It restructures the directery structure to look like:

```
backend/
    - .env
    - .gitignore
    - ...
    - static/
    - prompting/
    - testing/
        - testfiles...
    - app/
        - __init__.py
        - app.py
        - files...
```

This is needed to finish #22

See merge request sondralf/itp2-reflection-tool!18

* doc(back): command for running backend tests on windows

* doc: Folder structure

* linting

* test(back): Tests

* test(back): testing new endpoints

* testing creating courses

* test(back): Test user admin and not admin

* lint

* test wrong pipeline

* pipeline

* pipeline

* pipeline

* lint

* pipeline

* pipeline fix

* linting

* fix pipeline

* new pipeline

* feat(back): Pipeline backend tests

* test main

* Linting

* feat(back): Adds tests for backend userendpoint

* linting

* ref(back): Moves the files and create tests

Co-authored-by: Sondre Alfnes <sondre.alfnes@gmail.com>
Co-authored-by: Julian <julianao@stud.ntnu.no>