feat(back): Check if user is a teacher (!21)
When creating a account, the person is automaticly set to False as a Admin.

This merge requests adds a check to see if the user is a admin when a account is created, and if so, it will set the user to True as a Admin.

This will happen if a user has the role "Teacher"

Closes #32

See merge request sondralf/itp2-reflection-tool!21

* lint

* fix

* feat(back): Registrer as admin if a user is a lecturer

* UUpdate postman

* feat(back): Check if user is a teacher

Co-authored-by: Sondre Alfnes <sondre.alfnes@gmail.com>