ChatGPT prompt (!30)
To test this, create an environment variable in Postman and call it "api_key". Insert the OpenAI key value here.

There is a JSON file inside "prompting/data" called "inputToPostman". Copy this content, and use this as the body for the "analyze_feedback" endpoint

The current version of the prompting have been approved by PO as more than sufficient.

Co-authored-by: @jakobop 

Closes #24


Co-authored-by: Sondre Alfnes <sondre.alfnes@gmail.com>
Co-authored-by: Jakob Opland <jakobop@stud.ntnu.no>
Co-authored-by: Julian <julianao@stud.ntnu.no>