feat(back): Return units to a course (!61)
Add more data to the user endpoint. The user does now give the unit ids connected to each enrollment togheter with its date.

To test this run the /user endpoint.
This is implemented so that the user can get a visual representation of how many units are unanswered to a unit.
For overview info page. 


Note:
- It also filters out:
   - units that are in the future 
   - units that are answered


![Screenshot_2024-04-05_at_13.37.23](/uploads/5f15c1769c62d9172fcdb6d2b39897b2/Screenshot_2024-04-05_at_13.37.23.png)


Co-authored-by: Sondre Alfnes <sondre.alfnes@gmail.com>