feat(front): reflections since last report in unit card (#78) (!76)
Added column in unit table "reflections_since_last_report" and used this in frontend.

Also removed unused pages in frontend

You can test this by opening the overview of all the units, as a lecturer and see that the unavailable uniti has a "not available yet badge". If you submit a reflection on one of the units as a student, when viewing the unit card as a lecturer you should see a "Ready for report generation". If you then generate a report and then submit a report as another student, it should say "+1 reflection since last report"

Closes #78


Co-authored-by: Sondre Alfnes <sondralf@stud.ntnu.no>
Co-authored-by: Espen Iversen <espeiv@stud.ntnu.no>
Co-authored-by: Jakob Opland <jakobop@stud.ntnu.no>
Co-authored-by: Otto Hofseth Fearnley <ottohf@stud.ntnu.no>